refactor: simiplify code by fusing some of the leaderboard components
